// src/generated/models/Dev_tableassignrolesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dev_tableassignrolesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dev_tableassignrolesstatecode = keyof typeof Dev_tableassignrolesstatecode;
export const Dev_tableassignrolesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dev_tableassignrolesstatuscode = keyof typeof Dev_tableassignrolesstatuscode;

export interface Dev_tableassignrolesBase {
  dev_person?: string;
  dev_tableassignroleid: string;
  "dev_tablerole@odata.bind"?: string;
  "dev_tablevirtualizationprocess@odata.bind"?: string;
  "dev_user@odata.bind"?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dev_tableassignrolesstatecode;
  statuscode?: Dev_tableassignrolesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dev_tableassignroles extends Dev_tableassignrolesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dev_tablerolename?: string;
  dev_tablevirtualizationprocessname?: string;
  dev_username?: string;
  dev_useryominame?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dev_tablerole?: object;
  _dev_tablerole_value?: string;
  dev_tablevirtualizationprocess?: object;
  _dev_tablevirtualizationprocess_value?: string;
  dev_user?: object;
  _dev_user_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}